+++ tests/testsqlite.nim
@@ -5,12 +5,19 @@ import os, strutils, sequtils, times
 import norm / sqlite
 
 
-db("test.db", "", "", ""):
+const
+  dbName = "test.db"
+  customDbName = "custom_test.db"
+
+
+db(dbName, "", "", ""):
   type
     User {.table: "users".} = object
       email {.unique.}: string
+      ssn: Option[int]
       birthDate {.
         dbType: "INTEGER",
+        notNull,
         parseIt: it.i.fromUnix().local(),
         formatIt: dbValue(it.toTime().toUnix())
       .}: DateTime
@@ -29,6 +36,7 @@ db("test.db", "", "", ""):
       book {.
         dbCol: "bookId",
         dbType: "INTEGER",
+        notNull,
         fk: Book
         parser: getBookById,
         formatIt: dbValue(it.id),
@@ -42,7 +50,7 @@ suite "Creating and dropping tables, CRUD":
 
       for i in 1..9:
         var
-          user = User(email: "test-$#@example.com" % $i,
+          user = User(email: "test-$#@example.com" % $i, ssn: some i,
                       birthDate: parse("200$1-0$1-0$1" % $i, "yyyy-MM-dd"))
           publisher = Publisher(title: "Publisher $#" % $i)
           book = Book(title: "Book $#" % $i, authorEmail: user.email,
@@ -65,20 +73,21 @@ suite "Creating and dropping tables, CRUD":
       let query = "PRAGMA table_info($#);"
 
       check dbConn.getAllRows(sql query % "users") == @[
-        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 1],
-        @[dbValue 1, dbValue "email", dbValue "TEXT", dbValue 0, dbValue nil, dbValue 0],
-        @[dbValue 2, dbValue "birthDate", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 0]
+        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 1],
+        @[dbValue 1, dbValue "email", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 2, dbValue "ssn", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 0],
+        @[dbValue 3, dbValue "birthDate", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 0]
       ]
       check dbConn.getAllRows(sql query % "books") == @[
-        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 1],
-        @[dbValue 1, dbValue "title", dbValue "TEXT", dbValue 0, dbValue nil, dbValue 0],
-        @[dbValue 2, dbValue "authorEmail", dbValue "TEXT", dbValue 0, dbValue nil, dbValue 0],
-        @[dbValue 3, dbValue "publisherTitle", dbValue "TEXT", dbValue 0, dbValue nil, dbValue 0],
+        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 1],
+        @[dbValue 1, dbValue "title", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 2, dbValue "authorEmail", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 3, dbValue "publisherTitle", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
       ]
       check dbConn.getAllRows(sql query % "editions") == @[
-        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 1],
-        @[dbValue 1, dbValue "title", dbValue "TEXT", dbValue 0, dbValue nil, dbValue 0],
-        @[dbValue 2, dbValue "bookId", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 0]
+        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 1],
+        @[dbValue 1, dbValue "title", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 2, dbValue "bookId", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 0]
       ]
 
   test "Create records":
@@ -214,4 +223,40 @@ suite "Creating and dropping tables, CRUD":
         dbConn.exec sql "SELECT NULL FROM books"
         dbConn.exec sql "SELECT NULL FROM editions"
 
-  removeFile "test.db"
+  test "Custom DB":
+    withCustomDb(customDbName, "", "", ""):
+      createTables(force=true)
+
+    withCustomDb(customDbName, "", "", ""):
+      let query = "PRAGMA table_info($#);"
+
+      check dbConn.getAllRows(sql query % "users") == @[
+        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 1],
+        @[dbValue 1, dbValue "email", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 2, dbValue "ssn", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 0],
+        @[dbValue 3, dbValue "birthDate", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 0]
+      ]
+      check dbConn.getAllRows(sql query % "books") == @[
+        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 1],
+        @[dbValue 1, dbValue "title", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 2, dbValue "authorEmail", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 3, dbValue "publisherTitle", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+      ]
+      check dbConn.getAllRows(sql query % "editions") == @[
+        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 1],
+        @[dbValue 1, dbValue "title", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 2, dbValue "bookId", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 0]
+      ]
+
+    withCustomDb(customDbName, "", "", ""):
+      dropTables()
+
+      expect DbError:
+        dbConn.exec sql "SELECT NULL FROM users"
+        dbConn.exec sql "SELECT NULL FROM publishers"
+        dbConn.exec sql "SELECT NULL FROM books"
+        dbConn.exec sql "SELECT NULL FROM editions"
+
+    removeFile customDbName
+
+  removeFile dbName